dg00x-runtime: code refactoring for runtime structure
